requirements.txt
amqp==5.3.1
asgiref==3.9.1
async-timeout==5.0.1
billiard==4.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2.18
django-ckeditor-5==0.2.18
elastic-transport==8.17.1
elasticsearch==8.13.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.21.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kombu==5.5.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==1.24.4
packaging==25.0
pillow==10.2.0
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.19.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.40.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
